LayoutTests/http/tests/security/xss-DENIED-regular-propterty-with-iframe-proto-expected.txt


Tests that making other frame window a prototype doesn't expose that window properties
PASS: innerHeight === originalInnerHeight should be 'true' and is.
PASS: this.innerHeight === originalInnerHeight should be 'true' and is.
